Fixed bug 1258354: msys autoconf fails to detect strncmpi Fixed bug 1258355: autoconf Makefile.src assumes UNIX Fixed bug 1258357: configure doesn't set TTY_GRAPHICS under mingw32 Fixed bug 1258358: No autoconf support for key handlers
Autoconf improvements:
	Better determination of default windowing system
	Improved detection of libpng/zlib
	Basic support for Microsoft Window's native graphics
	Support aclocal 1.7 (needed for msys development)
	Support "make depend"

+++ sys/autoconf/bootstrap.sh
@@ -13,8 +13,10 @@ rm -f autoconf.hin
 aclocal
 # We can't avoid running autotools in multiple directories,
 # but at least we can have only one cache.
-rm -rf ../../autom4te.cache
-mv autom4te.cache ../..
+if [ -d autom4te.cache ]; then		# aclocal v1.7 doesn't create cache
+    rm -rf ../../autom4te.cache
+    mv autom4te.cache ../..
+fi
 # autoheader must be run from topdir since configure.ac refers
 # to sys/autoconf/autoconf.hin
 (cd ../.. && autoheader -I sys/autoconf sys/autoconf/configure.ac)
